Commits:
Modified the LogAnalyzer so that it can take a String, which is the name of a log file.

Added the numberOfAccesses method to your LogAnalyzer.

Add a busiestHour method.

Add a quietestHour method.

Add a busiestTwoHour method to the logAnalyzer - so from midnight-2am, 1am-3am, 2am-4am, ... 10pm midnight, 11pm-1am

Added methods to analyze